% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingMILEM_bag_cv.R
\name{testingMILEM_bag_cv}
\alias{testingMILEM_bag_cv}
\title{testingMILEM_GK_multik}
\usage{
testingMILEM_bag_cv(MILdatTst, trainRes)
}
\arguments{
\item{MILdatTst}{testing data set}

\item{trainRes}{Results results from trainging results}
}
\value{
YHat the empirical risk of the test
}
\description{
Compute the empirical risk of cv or testing for MIL with multiple instance
}
