% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingMILEM.R
\name{trainingMILEM}
\alias{trainingMILEM}
\title{trainingMILEM}
\usage{
trainingMILEM(MILdat, p, ki, tol, kerType, sigma, lambda, testPurpose)
}
\arguments{
\item{MILdat}{multiple instance learning data}

\item{p}{dimensions of instance covariance}

\item{ki}{number of instances in each bags}

\item{tol}{error tolerance for EM algorithm}

\item{kerType}{type of kerenl function, "RBF"(if choose) or "linear" (else choose, automatically)}

\item{sigma}{parameter for RBF kernel, also act as tuning parameter}

\item{lambda}{tuning parameter}

\item{testPurpose}{test purpose, for crossivalidation, the loss function is chosen as the phi loss;
for test (default value) , the loss function is chosen as the classification loss.}
}
\value{
trainRes
}
\description{
This is training function for multiple instance learning using EM algorithm
MILdat,tol,maxIter,kerType,sigma,lambda
}
